Delete pid_file after the process ends
Since the process should be exited after sending
an INT signal, the pid file can be deleted. That
makes it easy to add a simple status method,
which checks if the pid file is there.

Additionally the pid file will be deletet if an
SIGINT or SIGTERM arrives.

A future improvement could be the monitoring of
the process and delete the pid file if the process
is killed.